Feat: Updated userDTOResponse to include any user defined events. Feat: updated user services to work with new DTO. Feat: Added a getter for user id

+++ src/main/java/com/t/j/appointmentcalendar/Backend/user/UserAccountController.java
@@ -27,7 +27,7 @@ public class UserAccountController {
     // To get a specific user account
     @GetMapping("/{id}")
     public ResponseEntity<UserDTOResponse> getUser(@PathVariable int id) {
-        UserDTOResponse foundAcc = userServices.getSpefifiedUser(id);
+        UserDTOResponse foundAcc = userServices.getSpefifiedUser((long) id);
         return ResponseEntity.status(HttpStatus.FOUND).body(foundAcc);
     }
